+++ CHANGELOG.md
@@ -31,7 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Updated exclusion list to remove newly added `project-origin` & `mapyourgrid` as they aren't ESM tools.
+- Updated exclusion list to remove newly added `project-origin`, as it isn't an ESM tool.
+- Rebrand project: `open-esm-analysis` -> `openmod-tracker`.
 
 ## 2025-08-27